Use new link hook ordering
