+++ ffi/dart/lib/hook.dart
@@ -1,1 +1,1 @@
-export 'src/hook_helpers/link_helpers.dart';
+export 'src/hook_helpers/link_helpers.dart' show SymbolKeeper;
